es 6.6.0 support
